orchestrator/requirements.txt
requests
python-dotenv
mcp
playwright
